Make casting explicit and correct some whitespace issues

+++ src/main/java/uk/ac/cam/cl/dtg/isaac/api/managers/GameManager.java
@@ -442,7 +442,7 @@ public class GameManager {
         
         double boardPercentage = 0;
         for (Float questionPercentage : questionPercentages) {
-            boardPercentage += questionPercentage / questionPercentages.size();
+            boardPercentage += (double) questionPercentage / (double) questionPercentages.size();
         }
         gameboardDTO.setPercentageCompleted((int) Math.round(boardPercentage));
